# Optimized std.datetime.date.TimeOfDay.fromISOString